+++ config/config.go
@@ -1,6 +1,8 @@
 package config
 
 import (
+	"os"
+
 	"github.com/BurntSushi/toml"
 )
 
@@ -95,4 +97,19 @@ func LoadConfig(filePath string) (*Config, error) {
 		return nil, err
 	}
 	return &config, nil
+}
+
+// SaveConfig 保存配置到 TOML 配置文件
+func SaveConfig(filePath string, config *Config) error {
+	file, err := os.Create(filePath)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	encoder := toml.NewEncoder(file)
+	if err := encoder.Encode(config); err != nil {
+		return err
+	}
+	return nil
 }